hotfix(API Calls): Prevent redundant API calls refactor: Store results of API calls in local storage, reducing need for multiple calls
