+++ next.config.js
@@ -18,6 +18,7 @@ const nextConfig = {
       'cdn.myoutfits.biz',
       'www.comercialbenavides.net',
       'cdn.britannica.com',
+      'img.freepik.com',
     ],
   },
 };
